rdbms: optimistic, exception(-)

+++ rdbms/src/main/resources/init.sql
@@ -5,8 +5,8 @@
 create table account1
 (
     id bigserial constraint account_pk primary key,
-    amount int constraint positive_amount CHECK (amount >= 0),
-    version int
+    amount bigint constraint positive_amount CHECK (amount >= 0),
+    version bigint
 );
